Add Controller and new Entities

+++ NewsApi.DataAccess/Entities/CategoryEntity.cs
@@ -2,7 +2,7 @@
 
 public class CategoryEntity
 {
-    public long Id { get; init; }
+    public int Id { get; init; }
     public string Name { get; init; } = string.Empty;
     public ICollection<NewsEntity> News { get; init; } = [];
 }
